order_executor/requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.2
docker==7.0.0